feat(organizations): add new hooks for team member management
- Introduced `useMemberActiveTask`, `useMemberIdentity`, `useTeamMemberMutations`, and `useTeamMemberRoleActions` hooks to enhance team member management capabilities.
- Updated `use-team-member-card` to utilize the new hooks, promoting better separation of concerns and improving code maintainability.
- Removed deprecated logic from `use-team-member-card` to streamline functionality and improve clarity.
