+++ apps/web/core/hooks/organizations/index.ts
@@ -29,6 +29,10 @@ export * from './teams/use-invalidate-request-to-join';
 export * from './teams/use-team-invitations';
 export * from './teams/use-team-member';
 export * from './teams/use-team-member-card';
+export * from './teams/use-member-active-task';
+export * from './teams/use-member-identity';
+export * from './teams/use-team-member-mutations';
+export * from './teams/use-team-member-role-actions';
 export * from './teams/use-team-tasks';
 export * from './teams/use-team-tasks-query';
 export * from './teams/use-create-task';
